% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeneric.R
\name{bbo_benchmark}
\alias{bbo_benchmark}
\title{One stop solution to run a BBOB experiment.}
\usage{
bbo_benchmark(optimizer, algorithm_id, data_directory, dimensions = c(2, 3, 5,
  10, 20, 40), instances = c(1:5, 41:50), replications = 1L,
  budget = 1e+07, noisy = FALSE)
}
\arguments{
\item{optimizer}{[\code{function}]\cr Optimizer under test. The
exact function signature is \code{function(par, fun, lower, upper,
max_eval)}, were \code{par} is an initial search point, \code{fun}
is the function to minimize, \code{lower} and \code{upper} are the
lower and upper bounds of the box constrained search space and
\code{max_eval} is the maximum number of function evaluations the
optimizer should perform. It is safe to ignore the \code{par} if
your algorithm requires an initial population for example. You can
also use more function evaluations than specified in
\code{max_eval}. The setting is mainly provided for convenience
and as a last resort stopping criterion to limit the
runtime. Please see the Example section for an example of such a
wrapper function.}

\item{algorithm_id}{[\code{character}]\cr Short name of
algorithm. Should likely contain the package name containing the
optimization algorithm as well as the version used.}

\item{data_directory}{[\code{character}]\cr Base directory for
result files. If it does not start with a '/' (or drive letter on
Windows), relative to the current working directory.}

\item{dimensions}{[\code{integer}]\cr List of dimensions in which
to test the optimization algorithm. Default is fine for official
BBOB use.}

\item{instances}{[\code{integer}]\cr What instances (trial IDs) to
use for the test. Default is fine for official BBOB use.}

\item{replications}{[\code{integer}]\cr Number of replications to
perform. Defaults to 1 for BBOB 2012 compatibility but was 3 for
BBOB 2009. Default is fine for official BBOB use.}

\item{budget}{[\code{integer}]\cr Maximal number of function
evaluations per run. If the desired precision is not reached and
the budget is not used up, a random independent restart will be
performed. Defaults to 10000000. You may want to lower this to
save some runtime at the expense of solving fewer function
instances.}

\item{noisy}{[\code{logical}]\cr Wether to use the noisy test
functions (\code{noisy=TRUE}) or the noiseless test functions
(default, \code{noisy=FALSE}).}
}
\value{
Nothing. Run for the side effect of creating a BBOB data
directory that can be used with the Python post-processing code.
}
\description{
One stop solution to run a BBOB experiment.
}
\examples{
## This is a minimal example for the built in L-BFGS-B optimizer.
\dontrun{
my_optimizer <- function(par, fun, lower, upper, max_eval) {
  optim(par, fun, method="L-BFGS-B",
        lower=lower, upper=upper,
        control=list(maxit=max_eval))
}

budget <- 10000
bbo_benchmark(my_optimizer, "l-bfgs-b", "optim_l-bfgs-b",
               budget=10000)
}
}

