% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeneric.R
\name{bbo_timing}
\alias{bbo_timing}
\title{Run BBOB timing experiment}
\usage{
bbo_timing(optimizer, dimensions = c(2, 3, 5, 10, 20, 40))
}
\arguments{
\item{optimizer}{[\code{function}]\cr Optimizer under test. The
exact function signature is \code{function(par, fun, lower, upper,
max_eval)}, were \code{par} is an initial search point, \code{fun}
is the function to minimize, \code{lower} and \code{upper} are the
lower and upper bounds of the box constrained search space and
\code{max_eval} is the maximum number of function evaluations the
optimizer should perform. It is safe to ignore the \code{par} if
your algorithm requires an initial population for example. You can
also use more function evaluations than specified in
\code{max_eval} but this might increase the runtime of the timing
experiment unneccesarily.}

\item{dimensions}{[\code{integer}]\cr List of dimensions in which
to time the optimization algorithm. Default is fine for official
BBOB use.}
}
\value{
Nothing. Run for the side effect of creating a BBOB data
directory that can be used with the Python post-processing code.
}
\description{
Run BBOB timing experiment
}

