% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeneric.R
\name{bbob_setup_experiment}
\alias{bbob_end_experiment}
\alias{bbob_f_eval}
\alias{bbob_log_restart}
\alias{bbob_n_evaluations}
\alias{bbob_optimality_gap}
\alias{bbob_setup_experiment}
\title{Setup a single BBOB experiment for a given function, trial id and dimension.}
\usage{
bbob_setup_experiment(algorithm_id, datapath, function_id, instance_id,
  dimension)

bbob_end_experiment()

bbob_log_restart(reason)

bbob_f_eval(x)

bbob_n_evaluations()

bbob_optimality_gap()
}
\arguments{
\item{algorithm_id}{[\code{character}]\cr Algorithm ID.}

\item{datapath}{[\code{character}]\cr Base directory of
results. Note that this will append to already existing data in
the directory.}

\item{function_id}{[\code{integer}]\cr ID of function under
test. The noise free functions are numbered 1 to 24 and the noisy
function 101 to 124.}

\item{instance_id}{[\code{integer}]\cr Trial ID to use.}

\item{dimension}{[\code{integer}]\cr Dimensionality of parameter
space.}

\item{reason}{[\code{character}]\cr Reason for restart.}

\item{x}{[\code{numeric}]\cr Parameter setting at which to
evaluate the test function.}
}
\description{
These functions is part of the low-level BaBOB interface and not
meant for general use. Please see \code{\link{bbo_benchmark}} for
the recommended high-level interface.
}

