% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_search.R
\name{random_search}
\alias{random_search}
\title{Naive implementation of pure random search.}
\usage{
random_search(par, fn, lower, upper, max_eval)
}
\arguments{
\item{par}{Initial parameter combination (ignored).}

\item{fn}{Function to minimize.}

\item{lower}{Lower bounds of search space.}

\item{upper}{Upper bounds of search space.}

\item{max_eval}{Maximium number of function evaluations to perform.}
}
\value{
Best solution found.
}
\description{
Naive implementation of pure random search.
}

